requirements.txt
requests~=2.28.1
pillow~=10.3.0
selectolax~=0.3.11
selenium~=4.19.0
opencv-python~=4.9.0.80
lxml~=5.2.1
imutils~=0.5.4
numpy~=1.23.5
scikit-image~=0.23.2
aiohttp~=3.8.3
bs4~=0.0.2
beautifulsoup4~=4.12.3
stem~=1.8.2
openpyxl~=3.0.10
WMI~=1.5.1
matplotlib~=3.8.4
pandas~=1.5.2
PyAutoGUI~=0.9.54
pyperclip~=1.8.2